refactor: Inject EthereumSCApi via Config
