+++ state-chain/runtime/src/chainflip.rs
@@ -31,6 +31,7 @@ mod bitcoin_block_processor;
 #[macro_use]
 pub mod elections;
 pub mod bitcoin_elections;
+pub mod ethereum_sc_calls;
 pub mod solana_elections;
 pub mod vault_swaps;
